Move the box components into its own package
