+++ components/box/box.go
@@ -1,4 +1,4 @@
-package components
+package box
 
 import (
 	"strings"
